Refactor jwt auth headers and cookie
