+++ setup.cfg
@@ -55,7 +55,7 @@ per-file-ignores =
   dmr/exceptions.py: WPS202
   # Base jwt auth, its header transport, and the request helpers
   # all have to live here, so the old import paths keep working:
-  dmr/security/jwt/auth.py: WPS202
+  dmr/security/jwt/auth/header.py: WPS202
   # Compilation targets are a bit different:
   dmr/_compiled/*.py: WPS226, WPS231, WPS232, WPS402
   # Our test client has some autogenerated code:
