+++ tests/test_unit/test_security/test_jwt/test_jwt_cookie_auth.py
@@ -25,7 +25,7 @@ from dmr.security.jwt import (
     JWToken,
     request_jwt,
 )
-from dmr.security.jwt.auth import JWTAsyncAuth, JWTSyncAuth
+from dmr.security.jwt.auth.header import JWTAsyncAuth, JWTSyncAuth
 from dmr.test import DMRAsyncRequestFactory, DMRRequestFactory
 
 _LEEWAY: Final = 30  # seconds
@@ -46,14 +46,16 @@ def test_cookie_jwt_schema(
     """Ensures that security scheme is correct for cookie jwt auth."""
     instance = typ()
 
-    assert instance.security_schemes == snapshot({
-        'jwt': SecurityScheme(
-            type='apiKey',
-            description='JWT token auth via cookie',
-            name='access_token',
-            security_scheme_in='cookie',
-        ),
-    })
+    assert instance.security_schemes == snapshot(
+        {
+            'jwt': SecurityScheme(
+                type='apiKey',
+                description='JWT token auth via cookie',
+                name='access_token',
+                security_scheme_in='cookie',
+            ),
+        },
+    )
     assert instance.security_requirement == snapshot({'jwt': []})
 
 
@@ -65,14 +67,16 @@ def test_cookie_jwt_custom_schema(
     """Ensures that cookie and scheme names are customizable."""
     instance = typ(cookie_name='my-jwt', security_scheme_name='my-scheme')
 
-    assert instance.security_schemes == snapshot({
-        'my-scheme': SecurityScheme(
-            type='apiKey',
-            description='JWT token auth via cookie',
-            name='my-jwt',
-            security_scheme_in='cookie',
-        ),
-    })
+    assert instance.security_schemes == snapshot(
+        {
+            'my-scheme': SecurityScheme(
+                type='apiKey',
+                description='JWT token auth via cookie',
+                name='my-jwt',
+                security_scheme_in='cookie',
+            ),
+        },
+    )
     assert instance.security_requirement == snapshot({'my-scheme': []})
 
 
@@ -184,13 +188,15 @@ def test_sync_cookie_jwt_auth_csrf_enforced(
 
     assert isinstance(response, HttpResponse)
     assert response.status_code == HTTPStatus.FORBIDDEN
-    assert json.loads(response.content) == snapshot({
-        'detail': [
-            {
-                'msg': 'CSRF Failed: CSRF cookie not set.',
-            },
-        ],
-    })
+    assert json.loads(response.content) == snapshot(
+        {
+            'detail': [
+                {
+                    'msg': 'CSRF Failed: CSRF cookie not set.',
+                },
+            ],
+        },
+    )
 
 
 @pytest.mark.asyncio
@@ -221,13 +227,15 @@ async def test_async_cookie_jwt_auth_csrf_enforced(
 
     assert isinstance(response, HttpResponse)
     assert response.status_code == HTTPStatus.FORBIDDEN
-    assert json.loads(response.content) == snapshot({
-        'detail': [
-            {
-                'msg': 'CSRF Failed: CSRF cookie not set.',
-            },
-        ],
-    })
+    assert json.loads(response.content) == snapshot(
+        {
+            'detail': [
+                {
+                    'msg': 'CSRF Failed: CSRF cookie not set.',
+                },
+            ],
+        },
+    )
 
 
 @pytest.mark.django_db
